Saving dynamic nodes in db to survive manager restarts

+++ livestreaming/manager/node_controller.py
@@ -1,10 +1,11 @@
-from asyncio import Lock, create_task, gather, wait_for, CancelledError, sleep
+from asyncio import Lock, create_task, gather, wait_for, CancelledError, sleep, Event, get_running_loop
 from threading import Thread
 from typing import Dict, Type, List, Optional, TypeVar
 from urllib.parse import urlparse
 import time
 
 from livestreaming.web import ensure_url_does_not_end_with_slash
+from livestreaming.misc import TaskManager
 from .cloud import CombinedCloudAPI, NodeCreateError
 from .cloud.cloud_deployment import init_node, remove_ssh_key
 from .cloud import DeploymentStatus, VmStatus
@@ -13,17 +14,20 @@ from .cloud.definitions import CloudInstanceDefsController, OrderedInstanceDefin
 from .cloud.dns_api import DNSManager, get_dns_api_by_provider
 from . import  ManagerSettings, logger
 from .node_types import NodeTypeBase, ContentNode, EncoderNode, BrokerNode, StorageNode, DistributorNode
-
+from .db import Database
 
 Node_Type_T = TypeVar('Node_Type_T', bound=NodeTypeBase)
 
 
 class NodeController:
-    def __init__(self, manager_settings: ManagerSettings, definitions: CloudInstanceDefsController):
+    def __init__(self, manager_settings: ManagerSettings, definitions: CloudInstanceDefsController,
+                 db: Database):
         self.manager_settings: ManagerSettings = manager_settings
         self.definitions: CloudInstanceDefsController = definitions
+        self.db: Database = db
+
         self.api: CombinedCloudAPI = CombinedCloudAPI(definitions)
-        self.node_startup_lock: Lock = Lock()
+        self.node_startup_delete_lock: Lock = Lock()
         self.nodes: Dict[Type[NodeTypeBase], List[NodeTypeBase]] = {
             EncoderNode: [],
             ContentNode: [],
@@ -32,8 +36,10 @@ class NodeController:
         }
         self.broker_node: Optional[BrokerNode] = None
         self.server_by_name: Dict[str, Server] = {}
-        self.node_by_id: Dict[int, NodeTypeBase] = {}
+        self.node_by_id: Dict[int, NodeTypeBase] = {}  # TODO Do we still need this???
         self.dns_manager: Optional[DNSManager] = None
+        self.dyn_nodes_initialized = Event()
+
         if manager_settings.cloud_deployment:
             self.dns_manager = get_dns_api_by_provider(definitions)
 
@@ -42,13 +48,15 @@ class NodeController:
             await self.dns_manager.get_all_records() # fill record cache
 
         await self._init_static_nodes()
-        await self._handle_orphaned_nodes()
+        await self._init_existing_and_handle_orphaned_nodes()
+        task = create_task(self._handle_orphaned_servers_task())
+        TaskManager.fire_and_forget_task(task)
 
     async def get_pending_nodes(self, node_type: Type[Node_Type_T]) -> List[Node_Type_T]:
         """Get all nodes that are currently being created or initialized."""
         # Wait for the lock as there might be a node being started right now (in start_content_node) and we might not
         # know yet which instance definition will be used.
-        async with self.node_startup_lock:
+        async with self.node_startup_delete_lock:
             nodes = self.nodes[node_type]
             list = []
             for node in nodes:
@@ -99,54 +107,75 @@ class NodeController:
                 logger.info(f"Could not create node with definition {definition.section_name}")
                 pass
 
-    async def _dynamic_server_lifecycle(self, ordered_defs: OrderedInstanceDefinitionsList, node: NodeTypeBase):
+    async def _dynamic_server_lifecycle(self, ordered_defs: Optional[OrderedInstanceDefinitionsList],
+                                        node: NodeTypeBase):
         """A task that manages the lifecycle of a dynamic server. It stops the server on an error."""
         try:
-            async with self.node_startup_lock:
-                node.server = await self._start_dynamic_server(ordered_defs)
-
             if node.server is None:
-                logger.error("Could not create a new node")
-                # Node will be stopped/removed in finally clause.
-                return
+                # We need a server.
+                if ordered_defs is None:
+                    raise Exception("ordered_defs must not be None when there is no server")
+                deployment_time_start = time.time()
+                async with self.node_startup_delete_lock:
+                    node.server = await self._start_dynamic_server(ordered_defs)
 
-            await wait_for(self.api.wait_node_running(node.server), 60)
-            logger.info(f"Created node with definition {node.server.instance_definition.section_name}")
+                if node.server is None:
+                    logger.error("Could not create a new node")
+                    # Node will be stopped/removed in finally clause.
+                    return
 
-            tasks = []
-            if self.dns_manager:
-                tasks.append(node.server.set_domain_and_records(self.dns_manager))
-            tasks.append(init_node(node))
-            await gather(*tasks)
+                await wait_for(self.api.wait_node_running(node.server), 90)
+                logger.info(f"Created node with definition {node.server.instance_definition.section_name}")
 
-            node.base_url = "https://" + node.server.domain
+                tasks = []
+                if self.dns_manager:
+                    tasks.append(wait_for(node.server.set_domain_and_records(self.dns_manager), 90))
+                tasks.append(wait_for(init_node(node), 8 * 60))
+                await gather(*tasks)
+
+                deployment_duration = time.time() - deployment_time_start
+                logger.info(f"Ordering, booting and deployment of node {node.server.name} took "
+                            f"{deployment_duration:.2f}s")
+                node.base_url = "https://" + node.server.domain
+
+                await self.db.save_node(node)
 
             # Hand over to node watchdog and stay there until the node should be stopped.
             if node.server.deployment_status == DeploymentStatus.OPERATIONAL:
                 logger.info(f"New {type(node).__name__} is operational, host {node.server.host}, url {node.base_url}")
                 await node.watchdog()
-
+        except CancelledError:
+            pass
         except Exception as e:
-            if not isinstance(e, CancelledError):
-                logger.exception(f"Exception in dynamic server lifecycle task (server url {node.base_url})")
-
+            logger.exception(f"Exception in dynamic server lifecycle task (server url {node.base_url})")
+            node.shutdown = True  # We have no further use of the server.
         finally:
             # Stop server when this task ends.
-            await self._stop_dynamic_server(node)
-            if self.dns_manager and node.server and node.server.dns_record:
-                await self.dns_manager.remove_record(node.server.dns_record)
+            if node.shutdown:
+                async with self.node_startup_delete_lock:
+                    await self._stop_dynamic_server(node)
 
     def stop_node(self, node: NodeTypeBase):
         if node.lifecycle_task:
             node.lifecycle_task.cancel()
 
     async def _stop_dynamic_server(self, node: NodeTypeBase) -> None:
-        self.nodes[type(node)].remove(node)
-        self.node_by_id.pop(node.id)
+        if not self.node_startup_delete_lock.locked():
+            logger.error("_stop_dynamic_server should only be called when node_startup_delete_lock is hold")
+
+        try:
+            self.nodes[type(node)].remove(node)
+        except ValueError:
+            pass
+        self.node_by_id.pop(node.id, None)
         if node.server and isinstance(node.server, DynamicServer):
+            logger.info(f"Stopping server {node.server.name}")
+            self.server_by_name.pop(node.server.name, None)
             await self.api.delete_node(node.server)
-            self.server_by_name.pop(node.server.name)
+            if self.dns_manager and node.server.dns_record:
+                await self.dns_manager.remove_record(node.server.dns_record)
             await remove_ssh_key(node.server.host)
+        await self.db.delete_node(node)
 
     async def _init_static_nodes_of_type(self, node_type: Type[NodeTypeBase], urls: str):
         node_urls = list(map(str.strip, urls.split(',')))
@@ -159,13 +188,14 @@ class NodeController:
             node = node_type()
             node.base_url = ensure_url_does_not_end_with_slash(url)
             node.server = StaticServer(url_parsed.netloc, url_parsed.hostname)
-            if node_type is BrokerNode:
-                self.broker_node = node
-            else:
-                self.nodes[node_type].append(node)
-            self.node_by_id[node.id] = node
-            self.server_by_name[node.server.name] = node.server
-            node.lifecycle_task = create_task(self._static_server_lifecycle(node))
+            async with self.node_startup_delete_lock:
+                if node_type is BrokerNode:
+                    self.broker_node = node
+                else:
+                    self.nodes[node_type].append(node)
+                self.node_by_id[node.id] = node
+                self.server_by_name[node.server.name] = node.server
+                node.lifecycle_task = create_task(self._static_server_lifecycle(node))
 
     async def _init_static_nodes(self) -> None:
         if ',' in self.manager_settings.static_broker_node_base_url:
@@ -190,20 +220,71 @@ class NodeController:
                                  f"server {node.server.name})")
                 await sleep(20)
 
-    async def _handle_orphaned_nodes(self) -> None:
+    async def _handle_orphaned_servers_task(self) -> None:
+        """Periodically checks for orphaned cloud servers and stops them."""
         if not self.manager_settings.cloud_deployment:
             return
 
-        orphaned_nodes = await self.api.get_all_nodes()
-        for server in orphaned_nodes:
-            server.deployment_status = DeploymentStatus.ORPHANED
-            logger.warn(f"Found orphaned node: {server}")
-            if self.manager_settings.remove_orphaned_nodes:
-                self.api.delete_node(server)
+        while True:
+            try:
+                await sleep(10 * 60)
+
+                # Use a lock to avoid race conditions when a node is getting started and the name was not yet saved.
+                async with self.node_startup_delete_lock:
+                    servers = await self.get_all_dynamic_servers()
+                    for server in servers.values():
+                        if server.name not in self.server_by_name:
+                            server.deployment_status = DeploymentStatus.ORPHANED
+                            logger.warn(f"Found orphaned server: {server}")
+                            if self.manager_settings.remove_orphaned_nodes:
+                                await self.api.delete_node(server)
+            except CancelledError:
+                raise
+            except Exception:
+                logger.exception("Error in _handle_orphaned_servers_task")
+
+    async def get_all_dynamic_servers(self) -> Dict[str, DynamicServer]:
+        ret_servers = {}
+        servers = await self.api.get_all_nodes()
+        for server in servers:
+            # Only servers that belong to the namespace/prefix.
+            if server.name.startswith(self.manager_settings.dynamic_node_name_prefix):
+                existing_server_or_new = self.server_by_name.get(server.name, server)
+                ret_servers[existing_server_or_new.name] = existing_server_or_new
+        return ret_servers
+
+    async def _init_existing_and_handle_orphaned_nodes(self):
+        """Init existing nodes and remove all nodes that don't have a running server anymore."""
+        servers_by_name = await self.get_all_dynamic_servers()
+        nodes, orphaned_nodes = await self.db.get_nodes(servers_by_name, get_running_loop())
+
+        for node in nodes:
+            if not isinstance(node.server, DynamicServer):
+                logger.error(f"Expected DynamicServer in _init_existing_and_handle_orphaned_nodes")
+
+            async with self.node_startup_delete_lock:
+                if not node.server.is_operational():
+                    # Stop server and remove node.
+                    await self._stop_dynamic_server(node)
+                else:
+                    if self.dns_manager:
+                        # ensure domain records are set
+                        await node.server.set_domain_and_records(self.dns_manager)
+                    # start watchdog
+                    self.nodes[type(node)].append(node)
+                    self.node_by_id[node.id] = node
+                    self.server_by_name[node.server.name] = node.server
+                    node.lifecycle_task = create_task(self._dynamic_server_lifecycle(None, node))
+
+        for orphaned_node in orphaned_nodes:
+            logger.info(f"Found an orphaned node in database with id {orphaned_node.id} and "
+                        f"url {orphaned_node.base_url}. There was no corresponding dynamic server found. Delete it.")
+            await self.db.delete_node(orphaned_node)
 
 
 class CloudDeploymentDisabledError(Exception):
     pass
+
 
 class BrokerNodeNotUnique(Exception):
     pass
